+++ src/main/java/bowling/domain/Open.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.domain;
 
 public class Open implements State {
