+++ src/main/java/bowling/domain/Strike.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.domain;
 
 public class Strike implements State {
